The RunManager class has gotten quite complex and long, and we need to try to encapsulate functionality outside of it.  Remove the runDir and runId properties in favor of the Execution.execution_directory and Execution.execution_id properties.  Likewise use Execution.execution_uri instead of runManager.execURI.
In the constructor, remove the mlock() call for now and the persistent RunManager variable - protecting the RunManager instance from clears doesn't seem to be working completely.  Needs work.

We seem to be jumping through hoops with UUIDs and removing the 'urn:uuid:' prefix. Remove the prefix so we don't have to.