Now that Sessions are centrally managed, calls to calls to Configuration.loadConfig() tend to highlight how annoyingly often the call to set(config, 'authentication_token', '') produces the warning about authentication token security.
To make this more reasonable, add a Configuration.notify() function that is used in the above set() call.  If this is the *first* setting of the authentication token, then we display the warning.  Otherwise, subsequent calls don't display the warning message.

Update the SessionTest class to reflect this change.  Laso do some better error handling when parsing tokens in Session.

refs https://github.com/DataONEorg/sem-prov-design/issues/188