Fix bugs found during branch testing where the view() and publish() calls fail. We had two issues:
The format_id in the DataObject instance for objects downloaded via a mn.get() call wasn't converted to a char from a java.lang.String, causing display issues.

For DataONENode.get(), when the sysmeta does not have a file name set, we assign a random UUID, but then hadn't set that back in the sysmeta.  Blank filenames where causing display issues in view().

refs https://github.com/DataONEorg/sem-prov-design/issues/199