Changes to function convert2DomNode( ) in EMLCoverage class and TemporalCoverageNestedStruct class
For a struct array (n x 1), for example element singleDateTime

(1) When we loop over the struct array, create a node for current
keySet{i}.

(2) Recursively to process the nested struct under the current keySet{i}

(3) Add the current node (node) to its parent node (dm_node)

This morning, I add property initialization in the constructor
function. But it is not easy to remove empty element node because an
empty node is not valid in EML validation. So I do not initialize the
element rangeOfDates for the TemoralCoverageNestedStruct class. This
issue need more consideration.

Thanks.